wallet: extract out public facing MigrateLegacyToDescriptor functions
To wallet/migration. Also, update the includes in the call sites.

This patch should be reviewed with --color-moved=dimmed-zebra.
