+++ src/wallet/rpc/wallet.cpp
@@ -16,6 +16,7 @@
 #include <util/translation.h>
 #include <wallet/context.h>
 #include <wallet/export.h>
+#include <wallet/migration.h>
 #include <wallet/receive.h>
 #include <wallet/rpc/util.h>
 #include <wallet/wallet.h>
