+++ src/wallet/wallet.cpp
@@ -3751,381 +3751,6 @@ bool CWallet::CanGrindR() const
     return !IsWalletFlagSet(WALLET_FLAG_DISABLE_PRIVATE_KEYS);
 }
 
-// Returns wallet prefix for migration.
-// Used to name the backup file and newly created wallets.
-// E.g. a watch-only wallet is named "<prefix>_watchonly".
-static std::string MigrationPrefixName(CWallet& wallet)
-{
-    const std::string& name{wallet.GetName()};
-    return name.empty() ? "default_wallet" : name;
-}
-
-bool DoMigration(CWallet& wallet, WalletContext& context, bilingual_str& error, MigrationResult& res, const bool load_on_startup = true) EXCLUSIVE_LOCKS_REQUIRED(wallet.cs_wallet)
-{
-    AssertLockHeld(wallet.cs_wallet);
-
-    // Get all of the descriptors from the legacy wallet
-    std::optional<MigrationData> data = wallet.GetDescriptorsForLegacy(error);
-    if (data == std::nullopt) return false;
-
-    // Create the watchonly and solvable wallets if necessary
-    if (data->watch_descs.size() > 0 || data->solvable_descs.size() > 0) {
-        DatabaseOptions options;
-        options.require_existing = false;
-        options.require_create = true;
-        options.require_format = DatabaseFormat::SQLITE;
-
-        WalletContext empty_context;
-        empty_context.args = context.args;
-
-        // Make the wallets
-        options.create_flags = WALLET_FLAG_DISABLE_PRIVATE_KEYS | WALLET_FLAG_BLANK_WALLET | WALLET_FLAG_DESCRIPTORS;
-        if (wallet.IsWalletFlagSet(WALLET_FLAG_AVOID_REUSE)) {
-            options.create_flags |= WALLET_FLAG_AVOID_REUSE;
-        }
-        if (wallet.IsWalletFlagSet(WALLET_FLAG_KEY_ORIGIN_METADATA)) {
-            options.create_flags |= WALLET_FLAG_KEY_ORIGIN_METADATA;
-        }
-        if (data->watch_descs.size() > 0) {
-            wallet.WalletLogPrintf("Making a new watchonly wallet containing the watched scripts\n");
-
-            DatabaseStatus status;
-            std::vector<bilingual_str> warnings;
-            std::string wallet_name = MigrationPrefixName(wallet) + "_watchonly";
-            std::unique_ptr<WalletDatabase> database = MakeWalletDatabase(wallet_name, options, status, error);
-            if (!database) {
-                error = strprintf(_("Wallet file creation failed: %s"), error);
-                return false;
-            }
-
-            data->watchonly_wallet = CWallet::CreateNew(empty_context, wallet_name, std::move(database), options.create_flags, /*born_encrypted=*/false, error, warnings);
-            if (!data->watchonly_wallet) {
-                error = _("Error: Failed to create new watchonly wallet");
-                return false;
-            }
-            res.watchonly_wallet = data->watchonly_wallet;
-            LOCK(data->watchonly_wallet->cs_wallet);
-
-            // Parse the descriptors and add them to the new wallet
-            for (const auto& [desc_str, creation_time] : data->watch_descs) {
-                // Parse the descriptor
-                FlatSigningProvider keys;
-                std::string parse_err;
-                std::vector<std::unique_ptr<Descriptor>> descs = Parse(desc_str, keys, parse_err, /*require_checksum=*/ true);
-                assert(descs.size() == 1); // It shouldn't be possible to have the LegacyScriptPubKeyMan make an invalid descriptor or a multipath descriptors
-                assert(!descs.at(0)->IsRange()); // It shouldn't be possible to have LegacyScriptPubKeyMan make a ranged watchonly descriptor
-
-                // Add to the wallet
-                WalletDescriptor w_desc(std::move(descs.at(0)), creation_time, 0, 0, 0);
-                if (auto spkm_res = data->watchonly_wallet->AddWalletDescriptor(w_desc, keys, "", false); !spkm_res) {
-                    throw std::runtime_error(util::ErrorString(spkm_res).original);
-                }
-            }
-
-            // Add the wallet to settings
-            UpdateWalletSetting(*context.chain, wallet_name, load_on_startup, warnings);
-        }
-        if (data->solvable_descs.size() > 0) {
-            wallet.WalletLogPrintf("Making a new watchonly wallet containing the unwatched solvable scripts\n");
-
-            DatabaseStatus status;
-            std::vector<bilingual_str> warnings;
-            std::string wallet_name = MigrationPrefixName(wallet) + "_solvables";
-            std::unique_ptr<WalletDatabase> database = MakeWalletDatabase(wallet_name, options, status, error);
-            if (!database) {
-                error = strprintf(_("Wallet file creation failed: %s"), error);
-                return false;
-            }
-
-            data->solvable_wallet = CWallet::CreateNew(empty_context, wallet_name, std::move(database), options.create_flags, /*born_encrypted=*/false, error, warnings);
-            if (!data->solvable_wallet) {
-                error = _("Error: Failed to create new watchonly wallet");
-                return false;
-            }
-            res.solvables_wallet = data->solvable_wallet;
-            LOCK(data->solvable_wallet->cs_wallet);
-
-            // Parse the descriptors and add them to the new wallet
-            for (const auto& [desc_str, creation_time] : data->solvable_descs) {
-                // Parse the descriptor
-                FlatSigningProvider keys;
-                std::string parse_err;
-                std::vector<std::unique_ptr<Descriptor>> descs = Parse(desc_str, keys, parse_err, /*require_checksum=*/ true);
-                assert(descs.size() == 1); // It shouldn't be possible to have the LegacyScriptPubKeyMan make an invalid descriptor or a multipath descriptors
-                assert(!descs.at(0)->IsRange()); // It shouldn't be possible to have LegacyScriptPubKeyMan make a ranged watchonly descriptor
-
-                // Add to the wallet
-                WalletDescriptor w_desc(std::move(descs.at(0)), creation_time, 0, 0, 0);
-                if (auto spkm_res = data->solvable_wallet->AddWalletDescriptor(w_desc, keys, "", false); !spkm_res) {
-                    throw std::runtime_error(util::ErrorString(spkm_res).original);
-                }
-            }
-
-            // Add the wallet to settings
-            UpdateWalletSetting(*context.chain, wallet_name, load_on_startup, warnings);
-        }
-    }
-
-    // Add the descriptors to wallet, remove LegacyScriptPubKeyMan, and cleanup txs and address book data
-    return RunWithinTxn(wallet.GetDatabase(), /*process_desc=*/"apply migration process", [&](WalletBatch& batch) EXCLUSIVE_LOCKS_REQUIRED(wallet.cs_wallet){
-        if (auto res_migration = wallet.ApplyMigrationData(batch, *data); !res_migration) {
-            error = util::ErrorString(res_migration);
-            return false;
-        }
-        wallet.WalletLogPrintf("Wallet migration complete.\n");
-        return true;
-    });
-}
-
-util::Result<MigrationResult> MigrateLegacyToDescriptor(const std::string& wallet_name, const SecureString& passphrase, WalletContext& context, bool load_wallet)
-{
-    std::vector<bilingual_str> warnings;
-    bilingual_str error;
-
-    // The only kind of wallets that could be loaded are descriptor ones, which don't need to be migrated.
-    if (auto wallet = GetWallet(context, wallet_name)) {
-        assert(wallet->IsWalletFlagSet(WALLET_FLAG_DESCRIPTORS));
-        return util::Error{_("Error: This wallet is already a descriptor wallet")};
-    } else {
-        // Check if the wallet is BDB
-        const auto& wallet_path = GetWalletPath(wallet_name);
-        if (!wallet_path) {
-            return util::Error{util::ErrorString(wallet_path)};
-        }
-        if (!fs::exists(*wallet_path)) {
-            return util::Error{_("Error: Wallet does not exist")};
-        }
-        if (!IsBDBFile(BDBDataFile(*wallet_path))) {
-            return util::Error{_("Error: This wallet is already a descriptor wallet")};
-        }
-    }
-
-    // Load the wallet but only in the context of this function.
-    // No signals should be connected nor should anything else be aware of this wallet
-    WalletContext empty_context;
-    empty_context.args = context.args;
-    DatabaseOptions options;
-    options.require_existing = true;
-    options.require_format = DatabaseFormat::BERKELEY_RO;
-    DatabaseStatus status;
-    std::unique_ptr<WalletDatabase> database = MakeWalletDatabase(wallet_name, options, status, error);
-    if (!database) {
-        return util::Error{Untranslated("Wallet file verification failed.") + Untranslated(" ") + error};
-    }
-
-    // Make the local wallet
-    std::shared_ptr<CWallet> local_wallet = CWallet::LoadExisting(empty_context, wallet_name, std::move(database), error, warnings);
-    if (!local_wallet) {
-        return util::Error{Untranslated("Wallet loading failed.") + Untranslated(" ") + error};
-    }
-
-    return MigrateLegacyToDescriptor(std::move(local_wallet), passphrase, context, load_wallet);
-}
-
-util::Result<MigrationResult> MigrateLegacyToDescriptor(std::shared_ptr<CWallet> local_wallet, const SecureString& passphrase, WalletContext& context, bool load_wallet)
-{
-    MigrationResult res;
-    bilingual_str error;
-    std::vector<bilingual_str> warnings;
-
-    DatabaseOptions options;
-    options.require_existing = true;
-    DatabaseStatus status;
-
-    const std::string wallet_name = local_wallet->GetName();
-
-    // Before anything else, check if there is something to migrate.
-    if (local_wallet->IsWalletFlagSet(WALLET_FLAG_DESCRIPTORS)) {
-        return util::Error{_("Error: This wallet is already a descriptor wallet")};
-    }
-
-    // Make a backup of the DB in the wallet's directory with a unique filename
-    // using the wallet name and current timestamp. The backup filename is based
-    // on the name of the parent directory containing the wallet data in most
-    // cases, but in the case where the wallet name is a path to a data file,
-    // the name of the data file is used, and in the case where the wallet name
-    // is blank, "default_wallet" is used.
-    const std::string backup_prefix = wallet_name.empty() ? MigrationPrefixName(*local_wallet) : [&] {
-        // fs::weakly_canonical resolves relative specifiers and remove trailing slashes.
-        const auto legacy_wallet_path = fs::weakly_canonical(GetWalletDir() / fs::PathFromString(wallet_name));
-        return fs::PathToString(legacy_wallet_path.filename());
-    }();
-
-    fs::path backup_filename = fs::PathFromString(strprintf("%s_%d.legacy.bak", backup_prefix, GetTime()));
-    fs::path backup_path = fsbridge::AbsPathJoin(GetWalletDir(), backup_filename);
-    if (!local_wallet->BackupWallet(fs::PathToString(backup_path))) {
-        return util::Error{_("Error: Unable to make a backup of your wallet")};
-    }
-    res.backup_path = backup_path;
-
-    bool success = false;
-
-    // Unlock the wallet if needed
-    if (local_wallet->IsLocked() && !local_wallet->Unlock(passphrase)) {
-        if (passphrase.find('\0') == std::string::npos) {
-            return util::Error{Untranslated("Error: Wallet decryption failed, the wallet passphrase was not provided or was incorrect.")};
-        } else {
-            return util::Error{Untranslated("Error: Wallet decryption failed, the wallet passphrase entered was incorrect. "
-                                            "The passphrase contains a null character (ie - a zero byte). "
-                                            "If this passphrase was set with a version of this software prior to 25.0, "
-                                            "please try again with only the characters up to — but not including — "
-                                            "the first null character.")};
-        }
-    }
-
-    // Indicates whether the current wallet is empty after migration.
-    // Notes:
-    // When non-empty: the local wallet becomes the main spendable wallet.
-    // When empty: The local wallet is excluded from the result, as the
-    //             user does not expect an empty spendable wallet after
-    //             migrating only watch-only scripts.
-    bool empty_local_wallet = false;
-
-    {
-        LOCK(local_wallet->cs_wallet);
-        // First change to using SQLite
-        if (!local_wallet->MigrateToSQLite(error)) return util::Error{error};
-
-        // Do the migration of keys and scripts for non-empty wallets, and cleanup if it fails
-        if (HasLegacyRecords(*local_wallet)) {
-            success = DoMigration(*local_wallet, context, error, res, load_wallet);
-            // No scripts mean empty wallet after migration
-            empty_local_wallet = local_wallet->GetAllScriptPubKeyMans().empty();
-        } else {
-            // Make sure that descriptors flag is actually set
-            local_wallet->SetWalletFlag(WALLET_FLAG_DESCRIPTORS);
-            success = true;
-        }
-    }
-
-    // In case of loading failure, we need to remember the wallet files we have created to remove.
-    // A `set` is used as it may be populated with the same wallet directory paths multiple times,
-    // both before and after loading. This ensures the set is complete even if one of the wallets
-    // fails to load.
-    std::set<fs::path> wallet_files_to_remove;
-    std::set<fs::path> wallet_empty_dirs_to_remove;
-
-    // Helper to track wallet files and directories for cleanup on failure.
-    // Only directories of wallets created during migration (not the main wallet) are tracked.
-    auto track_for_cleanup = [&](const CWallet& wallet) {
-        const auto files = wallet.GetDatabase().Files();
-        wallet_files_to_remove.insert(files.begin(), files.end());
-        if (wallet.GetName() != wallet_name) {
-            // If this isn’t the main wallet, mark its directory for removal.
-            // This applies to the watch-only and solvable wallets.
-            // Wallets stored directly as files in the top-level directory
-            // (e.g. default unnamed wallets) don’t have a removable parent directory.
-            wallet_empty_dirs_to_remove.insert(fs::PathFromString(wallet.GetDatabase().Filename()).parent_path());
-        }
-    };
-
-
-    if (success) {
-        Assume(!res.wallet); // We will set it here.
-        // Check if the local wallet is empty after migration
-        if (empty_local_wallet) {
-            // This wallet has no records. We can safely remove it.
-            std::vector<fs::path> paths_to_remove = local_wallet->GetDatabase().Files();
-            local_wallet.reset();
-            for (const auto& path_to_remove : paths_to_remove) fs::remove(path_to_remove);
-        }
-
-        if (load_wallet) {
-            LogInfo("Loading new wallets after migration...\n");
-            /** We only override the load_on_startup setting in case the user explicitly said
-             * that he does not want to load the wallet, otherwise keep the old wallet configuration */
-        } else {
-            UpdateWalletSetting(*context.chain, wallet_name, /*load_on_startup=*/false, warnings);
-        }
-        // Migration successful, if load_wallet is set load all the migrated wallets.
-        bool main_wallet_set{false};
-        for (std::shared_ptr<CWallet>* wallet_ptr : {&local_wallet, &res.watchonly_wallet, &res.solvables_wallet}) {
-            if (success && *wallet_ptr) {
-                std::shared_ptr<CWallet>& wallet = *wallet_ptr;
-                // Track db path
-                track_for_cleanup(*wallet);
-                assert(wallet.use_count() == 1);
-                std::string wallet_name = wallet->GetName();
-                wallet.reset();
-                if (load_wallet) {
-                    wallet = LoadWallet(context, wallet_name, /*load_on_start=*/std::nullopt, options, status, error, warnings);
-                    if (!wallet) {
-                        LogError("Failed to load wallet '%s' after migration. Rolling back migration to preserve consistency. "
-                                 "Error cause: %s\n", wallet_name, error.original);
-                        success = false;
-                        break;
-                    }
-                }
-                // Set the first wallet as the main one.
-                // The loop order is intentional and must always start with the local wallet.
-                if (!main_wallet_set) {
-                    res.wallet_name = wallet_name;
-                    if (load_wallet) res.wallet = std::move(wallet);
-                    main_wallet_set = true;
-                }
-                if (wallet_ptr == &res.watchonly_wallet) {
-                    res.watchonly_wallet_name = wallet_name;
-                } else if (wallet_ptr == &res.solvables_wallet) {
-                    res.solvables_wallet_name = wallet_name;
-                }
-            }
-        }
-    }
-    if (!success) {
-        // Make list of wallets to cleanup
-        std::vector<std::shared_ptr<CWallet>> created_wallets;
-        if (local_wallet) created_wallets.push_back(std::move(local_wallet));
-        if (res.watchonly_wallet) created_wallets.push_back(std::move(res.watchonly_wallet));
-        if (res.solvables_wallet) created_wallets.push_back(std::move(res.solvables_wallet));
-
-        // Get the directories to remove after unloading
-        for (std::shared_ptr<CWallet>& wallet : created_wallets) {
-            track_for_cleanup(*wallet);
-        }
-
-        // Unload the wallets
-        for (std::shared_ptr<CWallet>& w : created_wallets) {
-            if (w->HaveChain()) {
-                // Unloading for wallets that were loaded for normal use
-                if (!RemoveWallet(context, w, /*load_on_start=*/false)) {
-                    error += _("\nUnable to cleanup failed migration");
-                    return util::Error{error};
-                }
-                WaitForDeleteWallet(std::move(w));
-            } else {
-                // Unloading for wallets in local context
-                assert(w.use_count() == 1);
-                w.reset();
-            }
-        }
-
-        // First, delete the db files we have created throughout this process and nothing else
-        for (const fs::path& file : wallet_files_to_remove) {
-            fs::remove(file);
-        }
-
-        // Second, delete the created wallet directories and nothing else. They must be empty at this point.
-        for (const fs::path& dir : wallet_empty_dirs_to_remove) {
-            Assume(fs::is_empty(dir));
-            fs::remove(dir);
-        }
-
-        // Restore the backup
-        // Convert the backup file to the wallet db file by renaming it and moving it into the wallet's directory.
-        bilingual_str restore_error;
-        const auto& ptr_wallet = RestoreWallet(context, backup_path, wallet_name, /*load_on_start=*/std::nullopt, status, restore_error, warnings, /*load_after_restore=*/false, /*allow_unnamed=*/true);
-        if (!restore_error.empty()) {
-            error += restore_error + _("\nUnable to restore backup of wallet.");
-            return util::Error{error};
-        }
-        // Verify that the legacy wallet is not loaded after restoring from the backup.
-        assert(!ptr_wallet);
-
-        return util::Error{error};
-    }
-    return res;
-}
-
 void CWallet::CacheNewScriptPubKeys(const std::set<CScript>& spks, ScriptPubKeyMan* spkm)
 {
     for (const auto& script : spks) {
