+++ src/wallet/wallet.h
@@ -1132,24 +1132,6 @@ public:
         }
     }
 };
-
-struct MigrationResult {
-    std::string wallet_name;
-    std::optional<std::string> watchonly_wallet_name;
-    std::optional<std::string> solvables_wallet_name;
-    std::shared_ptr<CWallet> wallet;
-    std::shared_ptr<CWallet> watchonly_wallet;
-    std::shared_ptr<CWallet> solvables_wallet;
-    fs::path backup_path;
-};
-
-//! Do all steps to migrate a legacy wallet to a descriptor wallet
-[[nodiscard]] util::Result<MigrationResult> MigrateLegacyToDescriptor(const std::string& wallet_name, const SecureString& passphrase, WalletContext& context, bool load_wallet = true);
-//! Requirement: The wallet provided to this function must be isolated, with no attachment to the node's context.
-[[nodiscard]] util::Result<MigrationResult> MigrateLegacyToDescriptor(std::shared_ptr<CWallet> local_wallet, const SecureString& passphrase, WalletContext& context, bool load_wallet = true);
-
-//! Determine the path that the wallet is stored in
-util::Result<fs::path> GetWalletPath(const std::string& name);
 } // namespace wallet
 
 #endif // BITCOIN_WALLET_WALLET_H
